refactor(code): move CLI commands into client package
