+++ libs/code/deepagents_code/client/commands/auth.py
