+++ libs/code/deepagents_code/main.py
@@ -1146,9 +1146,9 @@ def parse_args() -> argparse.Namespace:
         Parsed arguments namespace.
     """
     from deepagents_code._constants import DEFAULT_AGENT_NAME
-    from deepagents_code.auth_commands import setup_auth_parser
-    from deepagents_code.config_commands import setup_config_parser
-    from deepagents_code.mcp_commands import setup_mcp_parsers
+    from deepagents_code.client.commands.auth import setup_auth_parser
+    from deepagents_code.client.commands.config import setup_config_parser
+    from deepagents_code.client.commands.mcp import setup_mcp_parsers
     from deepagents_code.output import add_json_output_arg
     from deepagents_code.skills import setup_skills_parser
 
@@ -2600,12 +2600,12 @@ def cli_main() -> None:
         # and should fall through to the later handlers instead of raising here.
         command = getattr(args, "command", None)
         if command == "config":
-            from deepagents_code.config_commands import run_config_command
+            from deepagents_code.client.commands.config import run_config_command
 
             sys.exit(run_config_command(args))
 
         if command == "auth" and getattr(args, "auth_command", None) == "path":
-            from deepagents_code.auth_commands import run_auth_command
+            from deepagents_code.client.commands.auth import run_auth_command
 
             sys.exit(run_auth_command(args))
 
@@ -2615,7 +2615,7 @@ def cli_main() -> None:
             sys.exit(run_doctor_command(args))
 
         if command == "tools":
-            from deepagents_code.tools_commands import run_tools_command
+            from deepagents_code.client.commands.tools import run_tools_command
 
             sys.exit(run_tools_command(args))
 
@@ -2641,7 +2641,7 @@ def cli_main() -> None:
         from deepagents_code.config import console, settings
 
         if command == "auth":
-            from deepagents_code.auth_commands import run_auth_command
+            from deepagents_code.client.commands.auth import run_auth_command
 
             sys.exit(run_auth_command(args))
 
@@ -3383,7 +3383,10 @@ def cli_main() -> None:
 
             execute_skills_command(args)
         elif args.command == "mcp":
-            from deepagents_code.mcp_commands import run_mcp_config, run_mcp_login
+            from deepagents_code.client.commands.mcp import (
+                run_mcp_config,
+                run_mcp_login,
+            )
             from deepagents_code.ui import show_mcp_help
 
             if args.mcp_command == "login":
