+++ libs/code/tests/unit_tests/client/__init__.py
@@ -1,0 +1,1 @@
+"""Client-side unit tests."""
