+++ libs/code/tests/unit_tests/client/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Client command unit tests."""
